% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_angular_diff}
\alias{add_angular_diff}
\title{Add angular difference wrt reference direction to movement or eye tracker data frame}
\usage{
add_angular_diff(df, direction = c(0, 0, 1))
}
\arguments{
\item{df}{Movement or eye tracker or combined/resampled data frame. Expects either columns `rot_x`,
`rot_y`, `rot_z`, or columns `gaze_direction_x`, `gaze_direction_y`, `gaze_direction_z`}

\item{direction}{Reference gaze vector direction (default: forward)}
}
\value{
Input data frame with column `angular_diff` added.
}
\description{
Takes a movement or eye tracker data frame, derives rotation matrices from orientation data,
applies them to a reference direction vector, and calculates the row-by-row
angular difference for the vector rotation.
}
