% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_fixation2target}
\alias{add_fixation2target}
\title{Add target fixations to eyetracker data}
\usage{
add_fixation2target(df, target, targetname = NULL)
}
\arguments{
\item{df}{Eyetracker data frame. Expected column is `focus_object_raw`.}

\item{target}{Unity's internal target name (e.g. as returned by
`find_unity_threat_name`), or vector of target names}

\item{targetname}{Optional suffix for the new column name (default: `fixation`
without suffix)}
}
\value{
Eyetracker data frame with new column `fixation` or
`fixation_targetname` added
}
\description{
Returns the same dataframe with new column indicating 0 (no fixation) and 1
(fixation) over time.
}
