% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_gaze_elevation}
\alias{add_gaze_elevation}
\title{Add gaze elevation to (head) movement or eye tracker data frame}
\usage{
add_gaze_elevation(df, direction = c(0, 0, 1))
}
\arguments{
\item{df}{Movement or eye tracker or combined/resampled data frame. Expects either columns `rot_x`,
`rot_y`, `rot_z`, or columns `gaze_direction_x`, `gaze_direction_y`, `gaze_direction_z`}

\item{direction}{Reference gaze vector direction (default: forward)}
}
\value{
Original data frame with column `gaze_elevation` in degrees added.
}
\description{
Takes a movement or eye tracker data frame, derives rotation matrices from orientation data,
and adds the gaze elevation over the horizontal plane.
}
