% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_in_view}
\alias{add_in_view}
\title{Add column indicating if a stationary or moving target is in foveal view}
\usage{
add_in_view(
  eye_df,
  ref_pos,
  object_diameter = 0,
  foveal_angle = 1,
  direction = c(0, 0, 1)
)
}
\arguments{
\item{eye_df}{An eyetracker data frame}

\item{ref_pos}{A data frame with reference position as list or movement
tibble with columns `pos_x`, `pos_y`, `pos_z`}

\item{object_diameter}{Diameter of the object in m, default: 0.}

\item{direction}{Reference gaze direction, default: forward}

\item{foveal_field}{The foveal field in degrees, default: 1°.}
}
\value{
`ref_movement` with new columns `in_view`
}
\description{
Returns the same data frame with added column `in_view`.
}
