% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_orientation2target}
\alias{add_orientation2target}
\title{Add direction of the orientation of a reference object relative to a
stationary target (in 2D, XZ).}
\usage{
add_orientation2target(
  ref_movement,
  target_position,
  relative_to_first = FALSE
)
}
\arguments{
\item{ref_movement}{A data frame with reference movement, e.g. the head or
the waist movement. Movement file should have `pos_x`, `pos_y`, `pos_z`,
`rot_x`, `rot_y`, `rot_z` columns.}

\item{target_position}{A list containing `pos_x` `pos_y` `pos_z` target position}

\item{relative_to_first}{= FALSE}
}
\value{
`rev_movement` with new columns `target_diff` (raw angle relative
to target) and `target_ratio` (cosine of angle)
}
\description{
Returns the same data frame with added cos(angle) and
angle over time.
}
