% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_orientation2target2}
\alias{add_orientation2target2}
\title{Add direction of the orientation of a reference object relative to a
target movement (in 2D, XZ).}
\usage{
add_orientation2target2(
  ref_movement,
  target_movement,
  join_by = "time",
  relative_to_first = FALSE
)
}
\arguments{
\item{ref_movement}{A data frame with reference movement, e.g. the head or
the waist movement. Movement file should have `pos_x`, `pos_y`, `pos_z`,
`rot_x`, `rot_y`, `rot_z` columns.}

\item{target_movement}{Movement of point of interest (e.g threat movement)}

\item{join_by}{= "time" Column by which to join the two data frames (character)}

\item{relative_to_first}{= FALSE}
}
\value{
`ref_movement` with columns `time`, `target_diff` (raw angle relative
to target) and `target_ratio` (cosine of angle) added
}
\description{
Returns the reference data frame with cos(angle) and angle over time added.
}
