% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_rot}
\alias{add_rot}
\title{Add rotation matrix to movement or eye tracker data frame}
\usage{
add_rot(df, direction = c(0, 0, 1))
}
\arguments{
\item{df}{Movement or eye tracker or combined/resampled data frame. Expects either columns `rot_x`,
`rot_y`, `rot_z`, or columns `gaze_direction_x`, `gaze_direction_y`, `gaze_direction_z`}

\item{direction}{Reference gaze direction. A single vector. Default: forward}
}
\value{
Same df with rotation matrices added
}
\description{
Takes a movement or eye tracker data frame, and adds rotation matrices from orientation data.
For gaze data, rotation is computed wrt a reference vector, usually forward gaze.
}
