% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_rot_move}
\alias{add_rot_move}
\title{Helper function that takes a movement data frame and adds
a new column `R` with the rotation matrices.}
\usage{
add_rot_move(df)
}
\arguments{
\item{df}{Movement data frame. Expected columns: `rot_x`, `rot_y`, `rot_z`}
}
\value{
Same df with rotation matrices added
}
\description{
Helper function that takes a movement data frame and adds
a new column `R` with the rotation matrices.
}
