% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_speed}
\alias{add_speed}
\title{Add movement speed to movement data frame}
\usage{
add_speed(df)
}
\arguments{
\item{df}{Movement data frame. Expects columns `rot_x`, `rot_y`, `rot_z`,
`time`.}
}
\value{
Input data frame with column `speed` added.
}
\description{
Takes a movement data frame and calculates the row-by-row speed
}
