% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movement_functions.R
\name{add_vector2target}
\alias{add_vector2target}
\title{Add vector of a reference object relative to a
stationary target}
\usage{
add_vector2target(ref_movement, target_position)
}
\arguments{
\item{ref_movement}{A data frame with reference movement, e.g. the head or
 the waist movement. Movement file should have `pos_x`, `pos_y`, `pos_z`,
columns. Eye-tracker file should have `gaze_origin_x`,
 `gaze_origin_y`,`gaze_origin_z`columns}

\item{target_position}{A list containing `pos_x` `pos_y` `pos_z` target position}
}
\value{
`rev_movement` with new columns `vector2target_x` `vector2target_y`
`vector2target_z`
}
\description{
Returns the same data frame with added vector.
}
