% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{angle_between}
\alias{angle_between}
\title{This uses R base functions to calculate the angle between two vectors in degrees,
accounting for rounding errors to avoid NaNs.}
\usage{
angle_between(x, y)
}
\arguments{
\item{x}{Vector of length d}

\item{y}{Vector of length d}
}
\value{
Angle in degrees
}
\description{
This uses R base functions to calculate the angle between two vectors in degrees,
accounting for rounding errors to avoid NaNs.
}
