% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{angle_diff_deg}
\alias{angle_diff_deg}
\title{Calculate angle difference in degrees, respecting angles <0 or >360.}
\usage{
angle_diff_deg(target_ang, ref_ang)
}
\arguments{
\item{target_ang}{Angle}

\item{ref_ang}{Reference angle}
}
\value{
Difference in angles in degress
}
\description{
Calculate angle difference in degrees, respecting angles <0 or >360.
}
