% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{average_timeseries}
\alias{average_timeseries}
\title{Average a list of resampled time series (movement, fruit collection or
eyetracking) data frames}
\usage{
average_timeseries(df, .cols = !tidyselect::starts_with("time"))
}
\arguments{
\item{df}{A list of time series data frames. Expected column: 'new_time'}

\item{.cols}{A tidy selection of columns to be included in the output
(default: everything other than 'time')}
}
\value{
a data frame encapsulated in a list
}
\description{
Takes a list of resampled time series data frames and averages them into one
new data frame, by default removing a possible column 'time' if it exists. For
use with 'summarise'.
}
\details{
Expects a column 'new_time' as a reference.
No input checks are done - if the input data frames do not have the same size
or do not have the same 'new_time' index, no warning is thrown.
}
