% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{average_timeseries_cols}
\alias{average_timeseries_cols}
\title{Average time series data columns within group}
\usage{
average_timeseries_cols(
  df,
  .cols = tidyselect::ends_with("_movement_data_0_resampled"),
  ...
)
}
\arguments{
\item{df}{A trial results data frame.}

\item{.cols}{Columns containing timeseries dataframes to average
(Default: all resampled movement data). Each data frame must have a column
'new_time'.}

\item{...}{grouping variable(s), arguments separated with commas}
}
\value{
A new data frame with averaged timeseries data.
}
\description{
Convenient wrapper around `average_timeseries`. For provided columns,
averages resampled timeseries dataframes contained. Returns a new data frame
with the averaged data and columns with grouping variables. No input checks
are performed, i.e. if input data are not resampled at the same time index then
no warning is thrown.
}
