% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{calculate_2d_angle_deg}
\alias{calculate_2d_angle_deg}
\title{Calculate 2D angle between positive z-axis and a line from reference to target position in degrees}
\usage{
calculate_2d_angle_deg(target_x, target_z, ref_x = 0, ref_z = 0)
}
\arguments{
\item{target_x}{Vector of target x positions}

\item{target_z}{Vector of target y positions}

\item{ref_x}{Vector of reference x positions (default: origin)}

\item{ref_z}{Vector of reference y positions (default: origin)}
}
\value{
Angle between a line from reference to target vector and positive
z-axis, measured clockwise and in [-180, 180]
}
\description{
Calculate 2D angle between positive z-axis and a line from reference to target position in degrees
}
