% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{calculate_2d_dist}
\alias{calculate_2d_dist}
\title{Calculate 2D Euclidean distance between two positions}
\usage{
calculate_2d_dist(target_x, target_z, ref_x = 0, ref_z = 0)
}
\arguments{
\item{target_x}{Vector of target x positions}

\item{target_z}{Vector of target y positions}

\item{ref_x}{Vector of reference x positions (default: origin)}

\item{ref_z}{Vector of reference y positions (default: origin)}
}
\description{
Calculate 2D Euclidean distance between two positions
}
