% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{cart2sph}
\alias{cart2sph}
\title{Transform cartesian to spherical coordinates}
\usage{
cart2sph(x)
}
\arguments{
\item{x}{Vector of x/y/z coordinates}
}
\value{
Spherical coordinates, using pracma conventions
}
\description{
Wrapper around the pracma function with same name, handles NULL and NA arguments
}
