% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{check_tracker_cols}
\alias{check_tracker_cols}
\title{Check movement data columns}
\usage{
check_tracker_cols(df, start_pos, ref_tracker = "head_movement_data_0")
}
\arguments{
\item{df}{A trial results data frame with movement data columns}

\item{start_pos}{A trial start position column (expects, per row, a list with elements `x` and `z`) (string)}

\item{ref_tracker}{A reference tracker column (string, default is "head_movement_data_0")}
}
\value{
The same dataframe, with new columns `valid_` which contains detailed
 information for each tracker (list), and a column `valid_trackers` that combines
 information from all trackers (Boolean)
}
\description{
Checks validity of movement tracker data by using `check_valid_movement` on all
movement data columns contained in the data frame. The reference column is
checked against the starting position; all other columns are checked against the
reference column.
}
