% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{combine_movement}
\alias{combine_movement}
\title{Combine and resample two movement data frames}
\usage{
combine_movement(
  df1,
  df2,
  min_time = min(c(df1$time, df2$time)),
  max_time = max(c(df1$time, df2$time)),
  samplingrate = 10
)
}
\arguments{
\item{df1}{A dataframe of movement (must contain standard trajectory
columns, i.e. `"time"`, `"pos_x"`, `"pos_y"`, `"pos_z"`).}

\item{df2}{A dataframe of movement (must contain standard trajectory
columns, i.e. `"time"`, `"pos_x"`, `"pos_y"`, `"pos_z"`).}

\item{min_time}{Minimum time within the resampled movements (taken from `"time"`
column) to search.}

\item{max_time}{Maximum time within the resampled movements (taken from `"time"`
column) to search.}

\item{samplingrate}{resampling rate}
}
\value{
a combined data frame with original position columns suffixed by .1 and .2
}
\description{
This is a convenience function used in extract_ and get_ functions that deal
with two movement data frames.
}
\details{
NOTE: because sampling times can differ between data frames, data are
resampled at default rate of 10 Hz, and to avoid an impact of tracker
glitches they are median-smoothed over 3 data points (300 ms).
}
