% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_file_paths}
\alias{create_file_paths}
\title{Create Directory Paths Recursively}
\usage{
create_file_paths(base_path, sub_paths)
}
\arguments{
\item{base_path}{String. Root path to build from.}

\item{sub_paths}{Character vector of subdirectory names.}
}
\value{
Named list of full paths corresponding to each `sub_path`.
}
\description{
Ensures subdirectories exist within a base path, creating them if necessary.
}
\examples{
paths <- create_file_paths("data/output", c("plots", "tables"))
}
