% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{create_out_folders}
\alias{create_out_folders}
\title{Create output folders under an experiment name}
\usage{
create_out_folders(experiment_name)
}
\arguments{
\item{experiment_name}{A string representing the experiment name.}
}
\value{
Nothing.
}
\description{
Create output folders under an experiment name
}
