% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{create_resampling_index}
\alias{create_resampling_index}
\title{Create resampling index at desired sample rate}
\usage{
create_resampling_index(max_t, sample_rate)
}
\arguments{
\item{max_t}{Maximal time in seconds (numeric)}

\item{sample_rate}{Sample rate in Hz (numeric)}
}
\value{
A resampling index vector.
}
\description{
Create resampling index at desired sample rate
}
