% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{create_threat_name}
\alias{create_threat_name}
\title{Create threat name column from episode name}
\usage{
create_threat_name(df)
}
\arguments{
\item{df}{A trial results data frame.}
}
\value{
The data frame with the new column added.
}
\description{
Creates "threat" column, parsed from episode name. Expects column called
`episode`
For finding Unity's internal threat name, see "find_unity_threat_name"
}
