% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{cummean_na}
\alias{cummean_na}
\title{Calculate cummean with previous neighbour interpolation}
\usage{
cummean_na(x)
}
\arguments{
\item{x}{A vector of numeric}
}
\value{
A vector of numeric
}
\description{
Calculate cummean with previous neighbour interpolation
}
