% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{detect_sound}
\alias{detect_sound}
\title{Automatically detect sounds in audio recording using a volume/duration threshold}
\usage{
detect_sound(
  soundfile,
  start_time = 0,
  volume_threshold = 700,
  duration_threshold = 0.06
)
}
\arguments{
\item{soundfile}{A wave file name}

\item{start_time}{A start time (default 0)}

\item{volume_threshold}{A (vector of) volume treshold(s)}

\item{duration_threshold}{A (vector of) duration treshold(s) of same size}
}
\value{
a logical vector of same length as the threshold vectors
}
\description{
This function takes vectors of volume and duration thresholds and uses them to
detect above-threshold sounds. The duration threshold is evaluated cumulatively
across the entire recording (i.e. not necessarily contiguously). Default for
the threshold pair are optimised values from a labelled data set
}
