% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_fruit_task}
\alias{extract_fruit_task}
\title{Extract fruit picking between a `min_time` and `max_time`.
You can supply timestamps to `min_time` and `max_time` (e.g timestamp of
begin escape) to average over only a period. In this case, you can also
retrieve the collection rate by setting `rate` to 1.}
\usage{
extract_fruit_task(df, min_time = 0, max_time = Inf, method = "sum")
}
\arguments{
\item{df}{A fruit collection data frame. Expected columns are `time` and `event`.}

\item{min_time}{minimum time in s (default trial start)}

\item{max_time}{maximum time in s (default trial end)}

\item{method}{whether to extract sum (default) or rate}
}
\value{
Sum of fruits or average of fruits picked per second, between `min_time` and
  `max_time`.
}
\description{
NOTE: the fruit collection data frames contain no information on when
the trial ends; only on the last event of the task. Hence it is not possible
to extract the fruit collection rate without providing start and end time stamps.
}
