% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_movement_dist}
\alias{extract_movement_dist}
\title{Extract (min, max or mean) 2D distance between movement trajectory and
 reference position.}
\usage{
extract_movement_dist(
  ref_movement,
  ref_position,
  min_time = min(ref_movement$time),
  max_time = max(ref_movement$time),
  method = "min"
)
}
\arguments{
\item{ref_movement}{A dataframe of movement (must contain standard trajectory
columns, i.e. `"time"`, `"pos_x"`, `"pos_y"`, `"pos_z"`).}

\item{ref_position}{Reference position, stored as a list
containing `"pos_x"`, `"pos_z"` items.}

\item{min_time}{Minimum time within the ref_movement (taken from `"time"`
column) to search.}

\item{max_time}{Maximum time within the ref_movement (taken from `"time"`
column) to search.}

\item{method}{whether to extract min (default), max or mean}
}
\value{
A numeric value of the minimum 2D distance of the ref_movement
  trajectory to the supplied safe position.
}
\description{
#' Searches a movement trajectory dataframe and summarises distance
(ignoring y) to a reference position as average or minimum.
}
