% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_orientation}
\alias{extract_orientation}
\title{Extract direction of the orientation of a reference object relative to a
stationary target (in 2D, XZ) between a `min_time` and `max_time`.}
\usage{
extract_orientation(
  ref_movement,
  target_position,
  min_time = 0,
  max_time = Inf,
  convert2cos = TRUE,
  ...
)
}
\arguments{
\item{ref_movement}{A reference movement data frame, e.g. the head or
the waist movement. Movement file should have `pos_x`, `pos_y`, `pos_z`,
`rot_x`, `rot_y`, `rot_z` columns.}

\item{target_position}{A list with `pos_x`, `pos_y`, and `pos_z` coordinates of the target}

\item{min_time}{minimum time in s (default trial start)}

\item{max_time}{maximum time in s (default trial end)}

\item{convert2cos}{Whether to use the cosine of the angle (TRUE, default) or
the absolute angle (FALSE)}

\item{...}{any other arguments to be passed on to `add_orientation2target`}
}
\value{
An average of the `target_ratio` or absolute target difference over
 time between `min_time` and `max_time`. Uses trapezium rule for average over
 variable sample rate.
}
\description{
You can supply timestamps to `min_time` and `max_time` (e.g timestamp of
begin escape) to average over only a period.
}
