% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_recording_duration}
\alias{extract_recording_duration}
\title{Extract recording duration of a movement tibble}
\usage{
extract_recording_duration(df)
}
\arguments{
\item{df}{A movement tibble (must contain a column `time`)}
}
\value{
Scalar duration, i.e. difference between last and first time stamp
}
\description{
Extracting the recording duration of a movement tibble can be useful to check
the synchronisation of different trackers and devices.
}
