% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_speed}
\alias{extract_speed}
\title{Extract speed (peak or average speed) between a `min_time` and `max_time`.}
\usage{
extract_speed(
  df,
  min_time = min(df$time),
  max_time = max(df$time),
  method = "mean",
  samplingrate = 10
)
}
\arguments{
\item{df}{A movement data frame. Expected columns are `pos_x`, `pos_y`, `pos_z`}

\item{min_time}{minimum time in s (default: trial start)}

\item{max_time}{maximum time in s (default: trial end)}

\item{method}{whether to extract mean (default) or max}

\item{samplingrate}{resampling rate}
}
\value{
Peak or average 3-dimensional speed, between `min_time` and
  `max_time`.
}
\description{
You can supply timestamps to `min_time` and `max_time` (e.g timestamp of
begin escape) to average over only a period. In both cases, you can either
retrieve peak or average (default) three-dimensional speed.
}
\details{
NOTE: to avoid an impact of tracker glitches, data are resampled at default
rate of 10 Hz and median-smoothed over 3 data points (300 ms).
}
