% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_speed_angular}
\alias{extract_speed_angular}
\title{Extract angular speed (peak or average speed) for a direction vector
between a `min_time` and `max_time`.}
\usage{
extract_speed_angular(
  df,
  min_time = min(df$time),
  max_time = max(df$time),
  method = "mean",
  samplingrate = 10,
  direction = c(0, 0, 1)
)
}
\arguments{
\item{df}{A movement or eyetracker data frame. Expected columns are either
`rot_x`, `rot_y`, `rot_z`, or `gaze_direction_x`, `gaze_direction_y`,
`gaze_direction_z`}

\item{min_time}{minimum time in s (default: trial start)}

\item{max_time}{maximum time in s (default: trial end)}

\item{method}{whether to extract mean (default) or max}

\item{samplingrate}{resampling rate}

\item{direction}{Direction vector for which to compute the angular speed}
}
\value{
Peak or average 3-dimensional speed, between `min_time` and
  `max_time`.
}
\description{
You can supply timestamps to `min_time` and `max_time` (e.g timestamp of
begin escape) to average over only a period. In both cases, you can either
retrieve peak or average (default) angular speed.
}
\details{
NOTE: to avoid an impact of tracker glitches, data are resampled at default
rate of 10 Hz (movement) or 100 Hz (gaze) and median-smoothed over 3 data points
(300 or 30 ms).
}
