% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_functions.R
\name{extract_timestamp_duplicates}
\alias{extract_timestamp_duplicates}
\title{Extract duplicated time stamps.}
\usage{
extract_timestamp_duplicates(
  df,
  min_time = min(df$time),
  max_time = max(df$time),
  method = "max"
)
}
\arguments{
\item{df}{A summarised movement data frame (must contain a column `n`)}

\item{min_time}{Minimum time within the ref_movement (taken from `"time"`
column) to search.}

\item{max_time}{Maximum time within the ref_movement (taken from `"time"`
column) to search.}

\item{method}{whether to extract maximum number of duplicates (`max`, default),
total number of duplicates (`total`), or the number of duplicated time stamps
(`number`)}
}
\value{
Numeric, depending on `method`
}
\description{
Extracts the number of duplicated time stamps in a summarised movement
trajectory between a `min_time` and `max_time`.
Returns the number of duplicated time stamps, total number of duplicates,
or maximum number of duplicates.
}
