% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{find_argmax}
\alias{find_argmax}
\title{Find index of maximum, avoiding warnings and errors}
\usage{
find_argmax(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
The index of the minimum (a number or NA)
}
\description{
This is a wrapper for Rbase which.max() which avoids warnings and errors
when the supplied vector is empty or contains only NaNs
}
