% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{find_argmin}
\alias{find_argmin}
\title{Find index of minimum, avoiding warnings and errors}
\usage{
find_argmin(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
The index of the minimum (a number or NA)
}
\description{
This is a wrapper for Rbase which.min() which avoids warnings and errors
when the supplied vector is empty or contains only NaNs
}
