% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{find_fruit_position}
\alias{find_fruit_position}
\title{Find fruit position}
\usage{
find_fruit_position(scenario_data)
}
\arguments{
\item{scenario_data}{The scenario data list (should be read from scenario
.json file)}
}
\value{
A list representing the position of the Stand Here marker of the
  first found fruit task object in the scenario data. The list contains 6 items,
   `"pos_x"`, `"pos_y"`, and `"pos_z"`,
  the x, y, and z position of the object in Unity units (m), as well as  the
  Euler Angles  `"rot_x"`, `"rot_y"`, and `"rot_z"`. NULL if none found.
}
\description{
Finds the position of the `"Stand Here"` marker of the first found object
that matches the fruit task object names (`fruit_gameobject_names`). This
looks for child objects with name `"WalkHere"`.
}
