% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_hierarchy_object}
\alias{find_hierarchy_object}
\title{Find Hierarchy Object}
\usage{
find_hierarchy_object(x, name, include_children = FALSE, partial_match = FALSE)
}
\arguments{
\item{x}{The object hierarchy to search. This is typically a `list` loaded from a scenario `.json` file.}

\item{name}{The object name to search for.}

\item{include_children}{Should children be included in the result?}

\item{partial_match}{Should the name match exactly (FALSE, default) or can it be a part of the actual name?}
}
\value{
The matching object, or NULL if none found.
}
\description{
Search an object hierarchy for an object by name. This is useful for searching for a specific object in a saved scenario `.json` file.
}
