% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_hierarchy_object_any}
\alias{find_hierarchy_object_any}
\title{Find Hierarchy Object Matching Any Name}
\usage{
find_hierarchy_object_any(x, any_names, include_children = FALSE)
}
\arguments{
\item{x}{The object hierarchy to search. This is typically a `list` loaded
from a scenario `.json` file.}

\item{any_names}{A vector of object names to search for.}

\item{include_children}{Should children be included in the result?}
}
\value{
The matching object, or NULL if none found.
}
\description{
Extended version of [find_hierarchy_object], returning the first object that
matches any of a range of names.
}
