% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{find_initial_threat_position}
\alias{find_initial_threat_position}
\title{Find initial threat position from programmatic scenario planner}
\usage{
find_initial_threat_position(scenario_data)
}
\arguments{
\item{scenario_data}{The scenario data list (should be read from scenario
.json file)}
}
\description{
This refers to the front most part of the threat and does not correspond to
threat position in the threat movement data frame which is in reference to
the centre of the threat
}
