% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{find_max}
\alias{find_max}
\title{Find maximum, avoiding warnings}
\usage{
find_max(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
The maximum (a number or NA)
}
\description{
This is a wrapper for Rbase max() which avoids warnings when the supplied vector
is empty or contains only NaNs
}
