% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{find_min}
\alias{find_min}
\title{Find minimum, avoiding warnings}
\usage{
find_min(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
The minimum (a number or NA)
}
\description{
This is a wrapper for Rbase min() which avoids warnings when the supplied vector
is empty or contains only NaNs
}
