% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{find_safe_position}
\alias{find_safe_position}
\title{Find safe position}
\usage{
find_safe_position(scenario_data)
}
\arguments{
\item{scenario_data}{The scenario data list (should be read from scenario
.json file)}
}
\value{
A list representing the position of the first found Safe House object
  in the scenario data.  The list contains 6 items, `"pos_x"`, `"pos_y"`, and `"pos_z"`,
  the x, y, and z position of the object in Unity units (m), as well as  the
  Euler Angles  `"rot_x"`, `"rot_y"`, and `"rot_z"`. NULL if none found.
}
\description{
Finds the position of the Safe House. This looks objects with name
`"SafeHouse"`. If there are multiple, returns the first one found.
}
