% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{find_unity_threat_name}
\alias{find_unity_threat_name}
\title{Find Unity's internal threat name}
\usage{
find_unity_threat_name(scenario_data)
}
\arguments{
\item{scenario_data}{The scenario data list (should be read from scenario
.json file)}
}
\description{
Finds the name that Unity uses for the threat object. This is useful for
analysing eyetracking data.
}
