% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{fit_mdl}
\alias{fit_mdl}
\title{Fit (generalised) linear model, checking for convergence, and obtain marginal means}
\usage{
fit_mdl(DV, IV, df, family = "Gaussian")
}
\arguments{
\item{DV}{Dependent variable (string)}

\item{IV}{Independent variables (vector of string)}

\item{df}{Trial results data frame}

\item{family}{Error family (default: Gaussian)}
}
\value{
List with fitted fixed- and random-effects models, AIC difference
(negative in favour of random-effects), converging optimiser, estimated
marginal means
}
\description{
This function builds a model formula, evaluates FFX and RFX versions robustly,
and estimates marginal means from the model with higher model evidence (FFX or RFX)
}
