% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{fit_mdl_robust}
\alias{fit_mdl_robust}
\title{Fit (generalised) linear random effects model robustly, checking for convergence}
\usage{
fit_mdl_robust(mdl_formula, mdl_data, mdl_family = "Gaussian")
}
\arguments{
\item{mdl_formula}{Complete model formula}

\item{mdl_data}{Data frame}

\item{mdl_family}{Error family (default: Gaussian)}
}
\value{
List with fitted fixed- and random-effects models, AIC difference
(negative in favour of random-effects), converging optimiser, estimated
marginal means
}
\description{
This function fits a (G)LME. If the fit does not converge, it will try different
optimisers, and if still not converging will remove the integration over random effects.
}
