% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{fix_old_names}
\alias{fix_old_names}
\title{Fixes names in older versions of the output. Specifically, deletes the
occurance of `"vrthreat_"` in any column names.}
\usage{
fix_old_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Fixed data frame.
}
\description{
Fixes names in older versions of the output. Specifically, deletes the
occurance of `"vrthreat_"` in any column names.
}
