% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{fruit_gameobject_names}
\alias{fruit_gameobject_names}
\title{Common fruit object names used in VRthreat Unity}
\format{
An object of class \code{character} of length 5.
}
\usage{
fruit_gameobject_names
}
\description{
Common fruit object names used in VRthreat Unity
}
\keyword{datasets}
