% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{get_audio_section}
\alias{get_audio_section}
\title{Get audio section}
\usage{
get_audio_section(audiofile, from = 0, to = NA)
}
\arguments{
\item{audiofile}{A wave file}

\item{from}{Retrieve from this time point (in seconds after file start), default 0 (from file start)}

\item{to}{Retrieve to this time point  (in seconds after file start), default NA (to file end)}
}
\value{
Returns no value
}
\description{
Retrieves an audio recording from 'from' to 'to', using the tuneR library. Returns a wave object.
}
