% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{get_fruit_time}
\alias{get_fruit_time}
\title{Get fruit appearance or collect time for the n-th fruit
This function returns the exact time stamp of this fruit.}
\usage{
get_fruit_time(df, n = 1, event_type = "collect")
}
\arguments{
\item{df}{A fruit collection data frame. Expected columns are `time` and `event`.}

\item{n}{A specific fruit (integer) or "last" (default: 1)}

\item{event_type}{"appear" or "collect" (default)}
}
\value{
Time stamp
}
\description{
Get fruit appearance or collect time for the n-th fruit
This function returns the exact time stamp of this fruit.
}
