% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{get_time_of_first_event}
\alias{get_time_of_first_event}
\title{Get Time Of First Sequence Event}
\usage{
get_time_of_first_event(sequence_data, event_type)
}
\arguments{
\item{sequence_data}{The sequence data list (i.e. `sequence0_T00X.json`
parsed to a `list`).}

\item{event_type}{A vector containing the names event types. Should be
one of `"Wait"`, `"Event Call"`, `"Animate Behaviour"`, `"Move Towards`,
`"Rotate Towards"` `"Head Look"`. It can also be `"VerbaliseClip"`,
a specific animation of type `"Animate Behaviour"`. The time of the first
detected event is returned.}
}
\value{
A numeric value of the time (in Unity time, seconds since start-up)
  of the event. NA if none found.
}
\description{
Looks in the sequence data list to find the time at which the first
occurrence of a specified event name is called.
}
