% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{get_time_of_first_event_multi}
\alias{get_time_of_first_event_multi}
\title{Get Time Of First Sequence Event across multiple sequence columns}
\usage{
get_time_of_first_event_multi(df, event_type)
}
\arguments{
\item{df}{The entire trial_results data frame}

\item{event_type}{A string containing the name of the event type column. This
should contain a character vector of `"Wait"`, `"Event Call"`,
`"Animate Behaviour"`, `"Move Towards`, `"Rotate Towards"` `"Head Look"`.}
}
\value{
The entire trial_results data frame with columns `first_event_time`
and `first_event_index`  added
}
\description{
Looks in the specified rows find the time at which the first
occurrence of a specified event name is called, and takes the first across
all sequence columns
}
