% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_first_view}
\alias{guess_first_view}
\title{Guess first time an object is in foveal view}
\usage{
guess_first_view(eye_df, ref_pos, object_diameter = 0, foveal_angle = 1)
}
\arguments{
\item{eye_df}{An eyetracker data frame}

\item{ref_pos}{A data frame with reference position as list or movement
tibble with columns `pos_x`, `pos_y`, `pos_z`}

\item{object_diameter}{Diameter of the object in m, default: 0.}

\item{foveal_field}{The foveal field in degrees, default: 1°.}
}
\value{
A time stamp (real)
}
\description{
Guess first time an object is in foveal view
}
