% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_pos_at_time}
\alias{guess_pos_at_time}
\title{Guess/interpolate position at a point in time}
\usage{
guess_pos_at_time(df, ref_time)
}
\arguments{
\item{df}{a data frame}

\item{ref_time}{a reference time}
}
\value{
a list with elements `pos_x`, `pos_y`, `pos_z`
}
\description{
Guess/interpolate position at a point in time
}
