% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_threat_speed}
\alias{guess_threat_speed}
\title{Estimate actual threat speed from movement data}
\usage{
guess_threat_speed(threat_df, player_df, scenario_data, threat_appear_time)
}
\arguments{
\item{threat_df}{A movement tibble}

\item{player_df}{A movement tibble}

\item{scenario_data}{A scenario data structure}

\item{threat_appear_time}{Threat appearance time}
}
\value{
A list with threat_speed and move_time (the duration of the movement
        on which threat_speed is based on)
}
\description{
This function takes the threat movement up to player (2.5 m distance) or fruit
position (.5 m distance), whatever is earlier, and computes the threat speed
over this interval
}
