% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_time_at_dist2}
\alias{guess_time_at_dist2}
\title{Guess time at a distance between two movement data frames}
\usage{
guess_time_at_dist2(
  df1,
  df2,
  dist,
  min_time = min(c(df1$time, df2$time)),
  max_time = max(c(df1$time, df2$time)),
  method = "min",
  samplingrate = 10
)
}
\arguments{
\item{df1}{A dataframe of movement (must contain standard trajectory
columns, i.e. `"time"`, `"pos_x"`, `"pos_y"`, `"pos_z"`).}

\item{df2}{A dataframe of movement (must contain standard trajectory
columns, i.e. `"time"`, `"pos_x"`, `"pos_y"`, `"pos_z"`).}

\item{min_time}{Minimum time within the resampled movement (taken from `"time"`
column) to search.}

\item{max_time}{Maximum time within the resampled movement (taken from `"time"`
column) to search.}

\item{method}{whether to extract the first time distance is smaller ("min", default)
or larger than ("max") the threshold}

\item{samplingrate}{resampling rate}
}
\value{
A scalar time
}
\description{
Guess (by interpolation) the time that two movement data frames were, for the
first time, closer to (method "min") or further away than (method "max") a
distance threshold
}
