% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{orthdist}
\alias{orthdist}
\title{Perpendicular distance of p3 from p1 along p1p2}
\usage{
orthdist(p1, p2, p3)
}
\arguments{
\item{p1}{A vector of arbitrary dimension}

\item{p2}{Another vector with he same dimensions as p1}

\item{p3}{Yet another vector with the same dimensions}
}
\value{
A scalar
}
\description{
This can be used to determine when an object crosses a line
perpendicular to p1p2, for example along a line from threat to safety
}
