% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{play_audio_section}
\alias{play_audio_section}
\title{Play audio section}
\usage{
play_audio_section(audiofile, from = 0, to = NA)
}
\arguments{
\item{audiofile}{A wave file}

\item{from}{Play from this time point (in seconds after file start), default 0 (play from file start)}

\item{to}{Play to this time point  (in seconds after file start), default NA (play to file end)}
}
\value{
Returns no value
}
\description{
Plays an audio recording from 'from' to 'to', using the tuneR library and a native media player.
For Windows Media Player, player needs to be closed manually
tuneR needs write permissions, so a temporary file is copied to the current directory and deleted afterwards
This allows playing sounds from a write-protected directory but may make the function slow
}
