% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{prepare_gaze_data}
\alias{prepare_gaze_data}
\title{Prepare gaze data for further processing}
\usage{
prepare_gaze_data(df, samplingrate = NULL)
}
\arguments{
\item{df}{A movement or eyetracker data frame. Expected columns are either
`rot_x`, `rot_y`, `rot_z`, or `gaze_direction_x`, `gaze_direction_y`,
`gaze_direction_z`}

\item{samplingrate}{resampling rate}
}
\value{
Resampled data frame with filtered columns.
}
\description{
This is a convenience function to resample and filter head movement or
eye tracker data frames for further gaze processing.
}
\details{
To be used with `extract_...` functions that extract movement features between
time points. To avoid edge effects, the filtering is done on the entire data
frame.

NOTE: to avoid an impact of tracker glitches, data are resampled at default
rate of 10 Hz (movement) or 100 Hz (gaze) and median-smoothed over 3 data points
(300 or 30 ms).
}
