% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{proj_vec2plane}
\alias{proj_vec2plane}
\title{Project 3D-vector x into the plane spanned by y and z}
\usage{
proj_vec2plane(x, y, z)
}
\arguments{
\item{x}{A vector of length 3}

\item{y}{A vector of length 3}

\item{z}{Yet a vector of length 3}
}
\value{
Another vector of length 3
}
\description{
Project 3D-vector x into the plane spanned by y and z
}
