% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{proj_vec2vec}
\alias{proj_vec2vec}
\title{Project vector x onto vector y}
\usage{
proj_vec2vec(x, y)
}
\arguments{
\item{x}{A vector of any length}

\item{y}{A vector of the same length as x}
}
\value{
Another vector of the same length
}
\description{
Project vector x onto vector y
}
