% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_json_files}
\alias{read_json_files}
\title{Read JSON Files}
\usage{
read_json_files(data, .cols = ends_with("_json_location_0"), ...)
}
\arguments{
\item{data}{A dataframe of trial results}

\item{.cols}{A tidyselect specification of columns (e.g. `c(col1, col2)`).}

\item{...}{Additional arguments passed to `read_json`.
Default are columns that end with `_json_location_0`}
}
\description{
Read json files associated with trials from paths stored in supplied
columns. Expects absolute paths, use `update_file_pointers` to
update them if necessary. Function uses
`jsonlite::read_json` to read files.
}
