% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_log_messages}
\alias{read_log_messages}
\title{Read log messages}
\usage{
read_log_messages(df, messages, .colnames)
}
\arguments{
\item{df}{Data frame of trial results, containing column `res_path`}

\item{messages}{Vector of string messages to detect}

\item{.colnames}{Vector of column names to insert into `df`}
}
\value{
Original data frame with the following columsn added:
 `.colnames` (boolean), `.colnames_message`, `.colnames_time` ,
 `.colnames_number`, where the last three refer to the first detected
 message per trial
}
\description{
Read log messages from log files to combine with trial results
}
\examples{
# find "killed by magical force" messages in study 2
trials_with_log_messages <-
trials_raw \%>\%
   read_log_messages("ConfrontedThreat 0 You were killed by a magical force",
                     "killed_by_magical_force" = "...1")
}
