% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_participant_details}
\alias{read_participant_details}
\title{Read participant details from files}
\usage{
read_participant_details(df, data_dir)
}
\arguments{
\item{df}{A dataframe of trial results}

\item{data_dir}{Parent directory where data are stored}
}
\value{
Data frame with new columns
}
\description{
Will assume directory names assume to original participant ids, don''t use if
these have been changed.
}
