% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_functions.R
\name{read_quest_data}
\alias{read_quest_data}
\title{Read and score questionnaire data}
\usage{
read_quest_data(rawquestpath, metapath, raw = FALSE)
}
\arguments{
\item{rawquestpath}{The path that contains the questionnaires raw data files}

\item{metapath}{The path that contains the meta data. This must contain the
following 2 csv files: participants_excluded.csv, ppid_map.csv}

\item{raw}{(by default = FALSE) Determines the output. Read "Value" for more information.}
}
\value{
By default (when raw = FALSE): data frame containing the scored data of all the questionnaires.
Each row is a participant and each column is the score or subscore of a questionnaire

If raw = TRUE: data frame containing the raw data of all the questionnaires.
Each row is a participant and each column is the raw answer to a questionnaire item
}
\description{
Read and score questionnaire data
}
