% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_functions.R
\name{read_quest_raw}
\alias{read_quest_raw}
\title{Read raw (csv) questionnaire data}
\usage{
read_quest_raw(rawquestpath, prefix_fn, participants_excluded, redcap_ppid_map)
}
\arguments{
\item{rawquestpath}{The path that contains the questionnaires raw data files}

\item{prefix_fn}{Beginning of the questionnaire file name. Does not have to be exact as it
will be matched with the files in the rawquestpath.
3 csv files from Redcap exist:
- Ad
- Pr
- Po}

\item{participants_excluded}{list of excluded participants id (ppid)}

\item{redcap_ppid_map}{redcap id (survey_id) to VR id (ppid) map}
}
\value{
A data frame with the raw questionnaire data where each row is a participant and each column is a questionnaire item.
}
\description{
Read the questionnaire data csv files.
Removes the participants that didn't finish or that were excluded.
Removes unnecessary columns.
Maps ppid using fct replace_redcapid_ppid.
}
