% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_trial_results}
\alias{read_trial_results}
\title{Read Trial Results}
\usage{
read_trial_results(data_dir, ...)
}
\arguments{
\item{data_dir}{Parent directory where data are stored}

\item{...}{Arguments to be passed on to read_csv}
}
\value{
A combined data frame of all trial results.
}
\description{
Read & combine trial results files for a given experiment within a data
folder.
}
