% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{remove_tld}
\alias{remove_tld}
\title{Remove top-level directory in a full file path.}
\usage{
remove_tld(pathname)
}
\arguments{
\item{pathname}{A string path}
}
\value{
New pathname with top-level directory removed
}
\description{
Useful when top-level directory
name was changed during data curation
}
