% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{remove_tutorials}
\alias{remove_tutorials}
\title{Remove tutorial trials - i.e. episode either containing "Tutorial" or
"Practice"}
\usage{
remove_tutorials(df)
}
\arguments{
\item{df}{A trial results data frame.}
}
\value{
The data frame with the trials removed.
}
\description{
Remove tutorial trials - i.e. episode either containing "Tutorial" or
"Practice"
}
