% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_trial_functions.R
\name{resample_eyetracker_cols}
\alias{resample_eyetracker_cols}
\title{Resample eyetracker data columns}
\usage{
resample_eyetracker_cols(
  df,
  .cols = tidyselect::ends_with("_eye_tracking_data_0"),
  start_resample_col,
  end_resample_col,
  max_duration,
  sample_rate = 10
)
}
\arguments{
\item{df}{A trial results data frame.}

\item{.cols}{Columns containing dataframes to resample.
(Default: all eyetracker data). Each data frame must have a column 'time'.}

\item{start_resample_col}{Column containing start timestamps.}

\item{end_resample_col}{Column containing end timestamps.}

\item{max_duration}{Maximum duration of the resample.}

\item{sample_rate}{New sample rate (Hz).}
}
\value{
The data frame with a new columns of resampled movements added, with
new column name ending with `_resampled`.
}
\description{
Convenient wrapper around `resample_eyetracker`. For provided columns, resamples
eyetracker dataframes contained in the columns between times stored in
`start_resample_col` and `end_resample_col`. Resampling is done up to a max
duration of `max_duration` with a sample rate `sample_rate`. Non-existing
time points for any trial are filled with "NA".
Returns the data frame with new columns added.
}
