% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{resample_filter_pos}
\alias{resample_filter_pos}
\title{Resample and median-filter position and rotation columns in a movement data frame}
\usage{
resample_filter_pos(df, new_time, span = 3)
}
\arguments{
\item{df}{Movement data frame}

\item{new_time}{Resampling index}

\item{span}{Median filter span (default 3)}
}
\value{
Resampled data frame with filtered columns add as `new_pos_...` or `new_rot`
}
\description{
This is a convenience function to resample and filter an entire data frame.
To be used with `extract_...` functions that extract movement features between
time points. To avoide edge effects, the filtering is done on the entire data
frame, rather than just on an interval of interest.
}
