% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{rot2centralangle}
\alias{rot2centralangle}
\title{Compute central angle between two rotation matrices, for a direction vector}
\usage{
rot2centralangle(R1, R2, e)
}
\arguments{
\item{R1}{First rotation matrix}

\item{R2}{Second rotation matrix}

\item{e}{Base vector (e.g. c(0,0,1) for forward gaze), magnitude is ignored}
}
\value{
Angle in degrees
}
\description{
For a given direction vector, computes the central angle between the rotated
versions of the vector.
}
\details{
This is different from the distance between the rotations, because rotations
around the direction vector are ignored here. This is useful for computing
visual scan speed where rotation around the eye axis is irrelevant.
}
