% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{save_fig}
\alias{save_fig}
\title{Wrapper around `ggsave` to save figures.}
\usage{
save_fig(
  plt,
  experiment_name,
  fname,
  type = "cairo",
  device = "png",
  dpi = 600,
  width = 4,
  height = 4,
  ...
)
}
\arguments{
\item{plt}{The plot object.}

\item{experiment_name}{The experiment name.}

\item{fname}{The file name including extension.}

\item{type}{Type passed to `ggsave`}

\item{device}{Device passed to `ggsave` (default: "png")}

\item{dpi}{dpi passed to `ggsave`}

\item{width}{width of the plot (default units are inches)}

\item{height}{height of the plot (default units are inches)}

\item{...}{Other arguments passed to `ggsave`.}
}
\description{
Wrapper around `ggsave` to save figures.
}
