% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{scale_pastel}
\alias{scale_pastel}
\title{Pastel colored ggplot scale}
\usage{
scale_pastel(...)
}
\arguments{
\item{...}{Arguments passed to `scale_fill_hue` and `scale_color_hue`.}
}
\value{
A set of scales to be added to a ggplot.
}
\description{
A fill/color hue scale with pastel colours.
}
