% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{scale_vrthreat_endstate}
\alias{scale_vrthreat_endstate}
\title{VR Threat End State ggplot scale}
\usage{
scale_vrthreat_endstate(...)
}
\arguments{
\item{...}{Arguments passed to `scale_fill_manual` and `scale_color_manual`.}
}
\value{
A set of scales to be added to a ggplot.
}
\description{
A fill/color scale used across end state values (`"Safe"`,
`"ConfrontedThreat"`, `"Survived"`).
}
