% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{sound_classification_manual}
\alias{sound_classification_manual}
\title{Manually classify sounds}
\usage{
sound_classification_manual(
  df,
  trials = 1:nrow(df),
  sound_classification_file = "sound_classification.csv",
  sound_detection_file = "",
  from = "threat"
)
}
\arguments{
\item{df}{A trial results data frame}

\item{trials}{An integer selection of trials to play (corresponding to rows in output file), default: all rows}

\item{sound_classification_file}{An output file name (existing files will be re-used; otherwise a new file is created), default: `sound_classification.csv`}

\item{sound_detection_file}{A file with previously detected sounds (expected columns: `sound_detected` with 0/1 entries, `sound_comment`), default: use all sounds}

\item{from}{A string indicating whether to play sound from "start" or from "threat" (default)}
}
\description{
Play previously detected sounds from start or from threat appearance to end for manual
classification
}
