% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{sound_detection_manual}
\alias{sound_detection_manual}
\title{Manually detect sounds}
\usage{
sound_detection_manual(
  df,
  trials = NA,
  sound_detection_file = "sound_detection.csv",
  sound_classification_file = NA
)
}
\arguments{
\item{df}{A trial results data frame}

\item{trials}{An integer selection of trials to play (corresponding to rows in output file), default: all rows}

\item{sound_detection_file}{An output file name (existing files will be re-used; otherwise a new file is created), default: `sound_classification.csv`}

\item{sound_classification_file}{An output file name from an existing manual sound
classification. Only used if sound_detection_file does not exist. In this case, all trials with already detected sounds will be set to 1 rather than NA.}
}
\description{
Play all sounds of a study from start to end for manual
classification
}
