% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{split_path}
\alias{split_path}
\title{Split a path name into into its components}
\usage{
split_path(path)
}
\arguments{
\item{path}{A path name}
}
\value{
a vector of path components (starting with the last component)
}
\description{
Split a path name into into its components
}
