% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{standardise_path}
\alias{standardise_path}
\title{Standardise file path}
\usage{
standardise_path(path)
}
\arguments{
\item{path}{}
}
\value{
standardised path
}
\description{
This function ensures standardised path separators and capitalised drive 
letters. This is important for string operations on path definitions.
}
