% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_functions.R
\name{sub_sum_quest}
\alias{sub_sum_quest}
\title{Get sum of a questionnaire subscale}
\usage{
sub_sum_quest(qn, qnb, df)
}
\arguments{
\item{qn}{Name of the questionnaire}

\item{qnb}{Index of the items to include}

\item{df}{Data frame containing all the column names of the questionnaire data}
}
\value{
Sum of the specified questionnaire items
}
\description{
Get sum of a questionnaire subscale
}
