% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_functions.R
\name{sum_mssq}
\alias{sum_mssq}
\title{Get MSSQ subscale score (Motion Sickness; Golding, 2006)}
\usage{
sum_mssq(qn, df)
}
\arguments{
\item{qn}{MSSQ subscale name (mssq_a or mssq_c)}

\item{df}{Data frame containing all the column names of the questionnaire data}
}
\value{
Score of MSSQ subscale
}
\description{
Get MSSQ subscale score (Motion Sickness; Golding, 2006)
}
