% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_functions.R
\name{sum_quest}
\alias{sum_quest}
\title{Get sum of a questionnaire if questionnaire exists}
\usage{
sum_quest(qn, df)
}
\arguments{
\item{qn}{Name of the questionnaire}

\item{df}{Data frame containing all the column names of the questionnaire data}
}
\value{
Sum of all the items of that questionnaire
}
\description{
Get sum of a questionnaire if questionnaire exists
}
