% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{test_source_fun_run_tests_rep_coverage}
\alias{test_source_fun_run_tests_rep_coverage}
\title{Source Functions, Run Tests, and Report Coverage}
\usage{
test_source_fun_run_tests_rep_coverage(
  functions_dir,
  tests_dir,
  print_report = FALSE
)
}
\arguments{
\item{functions_dir}{The directory path containing function scripts.}

\item{tests_dir}{The directory path containing test scripts.}
}
\value{
Does not return a value; outputs coverage report.
}
\description{
This main function sources R scripts from a specified functions directory, runs
test files from a tests directory, and generates a test coverage report. It stops
execution if no function files are found and issues a warning if no test files are found.
}
\examples{
test_source_fun_run_tests_rep_coverage("relative_path/to/functions", "relative_path/to/tests")
}
