% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_vrthreat}
\alias{theme_vrthreat}
\title{VR Threat ggplot theme
#'}
\usage{
theme_vrthreat(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font familt}
}
\value{
A ggplot theme object.
}
\description{
VR Threat ggplot theme
#'
}
