% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{timeseries_mean}
\alias{timeseries_mean}
\title{Compute the mean across a vector `x` respecting time `t`, using trapezium
rule.}
\usage{
timeseries_mean(x, t)
}
\arguments{
\item{x}{Numeric values}

\item{t}{Time values}
}
\value{
Estimate of `x` across time values
}
\description{
Compute the mean across a vector `x` respecting time `t`, using trapezium
rule.
}
