% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{update_file_pointers}
\alias{update_file_pointers}
\title{Update file pointers}
\usage{
update_file_pointers(data, .cols = tidyselect::ends_with("_location_0"))
}
\arguments{
\item{data}{A dataframe of trial results}

\item{.cols}{A tidyselect specification of columns (e.g. `c(col1, col2)`).
Default are columns that end with `_location_0`}
}
\description{
Update pointers to CSV files associated with trials (e.g. Movement files)
in supplied columns. This updates the absolute paths (which might have been changed)
stored in columns provided combines with the column `res_path` as generated by
`read_trial_results`. This function takes care with changes in file location
or renaming of individual participant folders.
}
