% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{vec2rot}
\alias{vec2rot}
\title{Find a rotation matrix that maps one vector onto the other}
\usage{
vec2rot(x, y)
}
\arguments{
\item{x}{A vector in R3}

\item{y}{Another vector in R3}
}
\value{
a 3x3 rotation matrix R with y = Rx
}
\description{
Uses base R functions and pracma, to find the matrix that maps x to y by a rotation through
the normal to the plane spanned by x and y. This is the "shortest" mapping
between the two vectors, which does not involve any rotation about the rotated
vector itself.
}
\details{
This formula is independent of the coordinate system used, as the properties
of the output are defined by y = Rx and hold in any coordinate system.
This function is verified to be numerically precise for norm vectors
}
