% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrthreat-package.R
\docType{package}
\name{vrthreat-package}
\alias{vrthreat}
\alias{vrthreat-package}
\title{vrthreat: Analysis Tools for VRThreat Data}
\description{
Tools for analysing data from the Unity project "VRThreat". This package provides many generic functions useful for analysing any VR data using UXF format, and can handle all data generated by the public releases of the VRthreat Unity software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bachlab/vrthreat}
  \item Report bugs at \url{https://github.com/bachlab/vrthreat/issues}
}

}
\author{
\strong{Maintainer}: Dominik R Bach \email{d.bach@uni-bonn.de} (\href{https://orcid.org/0000-0003-3717-2036}{ORCID})

Authors:
\itemize{
  \item Jack Brookes (\href{https://orcid.org/0000-0002-4600-7561}{ORCID})
}

Other contributors:
\itemize{
  \item Juliana Sporrer (\href{https://orcid.org/0000-0002-9964-2876}{ORCID}) [contributor]
}

}
\keyword{internal}
