% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{write_audio_section}
\alias{write_audio_section}
\title{Write audio section}
\usage{
write_audio_section(audiofile, newfile, from = 0, to = NA)
}
\arguments{
\item{audiofile}{A wave file}

\item{newfile}{A new wave file}

\item{from}{Write from this time point (in seconds after file start), default 0 (from file start)}

\item{to}{Write to this time point  (in seconds after file start), default NA (to file end)}
}
\value{
Returns no value
}
\description{
Writes an audio recording from 'from' to 'to', using the tuneR library
}
