% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_log_message}
\alias{find_log_message}
\title{Find messages in a session log and assign to trials}
\usage{
find_log_message(df, messages, .colnames)
}
\arguments{
\item{df}{A session log data frame}

\item{messages}{A vector of string}

\item{.colnames}{A vector of column names, the same size as `messages`}
}
\value{
A data frame with `trial_num` and `.colnames` (boolean)
}
\description{
Find messages in a session log and assign to trials
}
