% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_log_message}
\alias{find_log_message}
\title{Find messages in a session log and assign to trials}
\usage{
find_log_message(df, messages)
}
\arguments{
\item{df}{A session log data frame}

\item{messages}{A vector of string}
}
\value{
A data frame with `trial_num` and `message` (boolean)
}
\description{
Find messages in a session log and assign to trials
}
