% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{get_first_fruit_collection}
\alias{get_first_fruit_collection}
\title{Get time of first fruit collection}
\usage{
get_first_fruit_collection(df)
}
\arguments{
\item{df}{Fruit collection data frame, expected columns: `time`, `event`}
}
\value{
a time stamp
}
\description{
Gets the time stamp of the first collected fruit from a fruit collection data frame
}
