% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_escape_end_time}
\alias{guess_escape_end_time}
\title{Guess end of escape time}
\usage{
guess_escape_end_time(
  safe_pos,
  end_time,
  end_state,
  ref_movement,
  buffer_time = 1.5,
  fade_time = 0.05
)
}
\arguments{
\item{safe_pos}{Safe position (output from find_safe_position)}

\item{end_time}{End time column entry}

\item{end_state}{End state column entry}

\item{ref_movement}{Reference movement (e.g. waist movement)}

\item{buffer_time}{Study-specific buffer time after trial end (default: 1.5 s)}

\item{fade_time}{Study-specific fade time (default: .05 s)}
}
\description{
Guess the time at which escape movement ends, depending on trial outcome:
Survived, killed: end of recording minus buffer and fade time
Escaped to safe house: crossing of the safe house threshold
Assumes safe house size 1 m x 1 m x 2 m (height)
If escaped but no safe house position logged, or entry to safe house unclear
(edge cases): end of recording minus 2 s (as a rough guess)
}
