% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{guess_move_from_pos}
\alias{guess_move_from_pos}
\title{Guess First Move From Position}
\usage{
guess_move_from_pos(
  pos,
  ref_movement,
  max_dist = 0.2,
  min_time = min(ref_movement$time),
  max_time = max(ref_movement$time)
)
}
\arguments{
\item{pos}{A list containing `x` `y` `z` of fixed position.}

\item{ref_movement}{A data frame with reference movement, e.g. the head or
the waist movement. Movement file should have `pos_x`, `pos_y`, `pos_z`
columns.}

\item{max_dist}{Maximum distance in meters before player is
classified as away from the position}
}
\value{
A time stamp (Unity time) representing the time when the participant
  initially moved from the position.
}
\description{
Guess the time at which the player initially moved away from a fixed position
}
