% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{play_audio_from_threat}
\alias{play_audio_from_threat}
\title{Play audio recording from threat appearance to the end}
\usage{
play_audio_from_threat(df)
}
\arguments{
\item{df}{A trial results data frame}
}
\value{
Returns no value
}
\description{
Finds threat start and location of microphone recording in a data frame, and plays the audio from the start of the threat, using 'play_audio_section'.
Expects columns 'threat_appear_time', 'start_time', 'mic.wav_location_0'. Expects absolute paths, use `update_file_pointers` to
update them if necessary.
}
