% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_log_messages}
\alias{read_log_messages}
\title{Read log messages}
\usage{
read_log_messages(data_dir, messages)
}
\arguments{
\item{data_dir}{Parent directory where data are stored}

\item{messages}{Vector of string messages to detect}
}
\description{
Read log messages from log files to combine with trial results
}
