% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_study_data}
\alias{read_study_data}
\title{Read complete study data}
\usage{
read_study_data(rawpath, metapath)
}
\arguments{
\item{rawpath}{The path that contains the individual subjects' directories.
The path name may have been changed during data curation, and the original
top-level directory name in the VR output will be replaced by this path.}

\item{metapath}{The path that contains the meta data. This must contain the
following csv files, which are used in their order of appeareance:
1. replacements.csv: joins interrupted participants. Must contain columns
   ppid_old, ppid_new, session_num_new, last_trial_exclude
2. participants_excluded.csv: participants to exclude. Must contain column
   ppid
3. trial_exclusions.csv: individual trials to exclude. Must contain columns
   ppid, session_num, trial_num}
}
\value{
A data frame with all experiment data
}
\description{
This function reads all VRthreat output data from one study into a data frame.
The
}
