% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_study_data}
\alias{read_study_data}
\title{Read complete study data}
\usage{
read_study_data(rawpath, metapath)
}
\arguments{
\item{rawpath}{The path that contains the individual subjects' directories.
The path name may have been changed during data curation, and the original
top-level directory name in the VR output will be replaced by this path.}

\item{metapath}{The path that contains the meta data. This must contain the
following 4 csv files: replacements.csv, participants_excluded.csv,
episodes_info.csv, threats_info.csv}
}
\value{
A data frame with all experiment data
}
\description{
This function reads all VRthreat output data from one study into a data frame.
The
}
