% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_audio_functions.R
\name{sound_classification_manual}
\alias{sound_classification_manual}
\title{Manually classify sounds}
\usage{
sound_classification_manual(
  df,
  audio_path,
  drop_tld = TRUE,
  trials = 1:nrow(df),
  sound_classification_file = "sound_classification.csv",
  sound_detection_file = ""
)
}
\arguments{
\item{df}{A trial results data frame}

\item{audio_path}{The path to the data directory}

\item{drop_tld}{Drop top-level directory in stored file location? (default TRUE)}

\item{trials}{An integer selection of trials to play (corresponding to rows in output file), default: all rows}

\item{sound_classification_file}{An output file name (existing files will be re-used; otherwise a new file is created), default: `sound_classification.csv`}

\item{sound_detection_file}{A file with previously detected sounds (expected columns: `sound_detected` with 0/1 entries, `sound_comment`), default: use all sounds}
}
\description{
Play previously detected sounds from threat appearance to end for manual
classification
}
