% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_movement_functions.R
\name{summarise_movement}
\alias{summarise_movement}
\title{Summarise movement data frame for each unique time stamp}
\usage{
summarise_movement(df, by)
}
\arguments{
\item{df}{Movement data frame}

\item{by}{= "time": time stamp key}
}
\description{
Ensures that each time stamp is unique by averaging all rows within duplicated time stamp
values (for rotation columns: circular mean). Also adds a column `n` which
is the count of rows for this time stamp.
This is useful because there can be several observations for a time stamp in a
movement data frame. One reason is that the epoch can be "frozen" when the
menu button is pressed, generating new data but freezing the timer.
}
